+++ tests/integrations/django/test_basic.py
@@ -823,236 +823,6 @@ def test_signals_spans_disabled(sentry_init, client, capture_events):
     assert not transaction["spans"]
 
 
-@pytest.mark.forked
-def test_cache_spans_middleware(sentry_init, client, settings, capture_events):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake",
-        }
-    }
-    settings.MIDDLEWARE.insert(0, "django.middleware.cache.UpdateCacheMiddleware")
-    settings.MIDDLEWARE.append("django.middleware.cache.FetchFromCacheMiddleware")
-
-    client.application.load_middleware()
-
-    sentry_init(
-        integrations=[
-            DjangoIntegration(
-                cache_spans=True,
-                middleware_spans=False,
-                signals_spans=False,
-            )
-        ],
-        traces_sample_rate=1.0,
-    )
-    events = capture_events()
-
-    client.get(reverse("not_cached_view"))
-    client.get(reverse("not_cached_view"))
-
-    (first_event, second_event) = events
-    assert len(first_event["spans"]) == 1
-    assert first_event["spans"][0]["op"] == "cache"
-    assert first_event["spans"][0]["description"].startswith(
-        "get views.decorators.cache.cache_header."
-    )
-    assert first_event["spans"][0]["data"] == {"cache.hit": False}
-
-    assert len(second_event["spans"]) == 2
-    assert second_event["spans"][0]["op"] == "cache"
-    assert second_event["spans"][0]["description"].startswith(
-        "get views.decorators.cache.cache_header."
-    )
-    assert second_event["spans"][0]["data"] == {"cache.hit": False}
-
-    assert second_event["spans"][1]["op"] == "cache"
-    assert second_event["spans"][1]["description"].startswith(
-        "get views.decorators.cache.cache_page."
-    )
-    assert second_event["spans"][1]["data"] == {
-        "cache.hit": True,
-        "cache.item_size": 58,
-    }
-
-
-@pytest.mark.forked
-def test_cache_spans_disabled_middleware(sentry_init, client, settings, capture_events):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake",
-        }
-    }
-    settings.MIDDLEWARE.insert(0, "django.middleware.cache.UpdateCacheMiddleware")
-    settings.MIDDLEWARE.append("django.middleware.cache.FetchFromCacheMiddleware")
-
-    client.application.load_middleware()
-
-    sentry_init(
-        integrations=[
-            DjangoIntegration(
-                cache_spans=False,
-                middleware_spans=False,
-                signals_spans=False,
-            )
-        ],
-        traces_sample_rate=1.0,
-    )
-    events = capture_events()
-
-    client.get(reverse("not_cached_view"))
-    client.get(reverse("not_cached_view"))
-
-    (first_event, second_event) = events
-    assert len(first_event["spans"]) == 0
-    assert len(second_event["spans"]) == 0
-
-
-def test_cache_spans_decorator(sentry_init, client, settings, capture_events):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake",
-        }
-    }
-    sentry_init(
-        integrations=[
-            DjangoIntegration(
-                cache_spans=True,
-                middleware_spans=False,
-                signals_spans=False,
-            )
-        ],
-        traces_sample_rate=1.0,
-    )
-    events = capture_events()
-
-    client.get(reverse("cached_view"))
-    client.get(reverse("cached_view"))
-
-    (first_event, second_event) = events
-    assert len(first_event["spans"]) == 1
-    assert first_event["spans"][0]["op"] == "cache"
-    assert first_event["spans"][0]["description"].startswith(
-        "get views.decorators.cache.cache_header."
-    )
-    assert first_event["spans"][0]["data"] == {"cache.hit": False}
-
-    assert len(second_event["spans"]) == 2
-    assert second_event["spans"][0]["op"] == "cache"
-    assert second_event["spans"][0]["description"].startswith(
-        "get views.decorators.cache.cache_header."
-    )
-    assert second_event["spans"][0]["data"] == {"cache.hit": False}
-
-    assert second_event["spans"][1]["op"] == "cache"
-    assert second_event["spans"][1]["description"].startswith(
-        "get views.decorators.cache.cache_page."
-    )
-    assert second_event["spans"][1]["data"] == {
-        "cache.hit": True,
-        "cache.item_size": 58,
-    }
-
-
-def test_cache_spans_disabled_decorator(sentry_init, client, settings, capture_events):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake",
-        }
-    }
-    sentry_init(
-        integrations=[
-            DjangoIntegration(
-                cache_spans=False,
-                middleware_spans=False,
-                signals_spans=False,
-            )
-        ],
-        traces_sample_rate=1.0,
-    )
-    events = capture_events()
-
-    client.get(reverse("cached_view"))
-    client.get(reverse("cached_view"))
-
-    (first_event, second_event) = events
-    assert len(first_event["spans"]) == 0
-    assert len(second_event["spans"]) == 0
-
-
-def test_cache_spans_templatetag(sentry_init, client, settings, capture_events):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake",
-        }
-    }
-    sentry_init(
-        integrations=[
-            DjangoIntegration(
-                cache_spans=True,
-                middleware_spans=False,
-                signals_spans=False,
-            )
-        ],
-        traces_sample_rate=1.0,
-    )
-    events = capture_events()
-
-    client.get(reverse("view_with_cached_template_fragment"))
-    client.get(reverse("view_with_cached_template_fragment"))
-
-    (first_event, second_event) = events
-    assert len(first_event["spans"]) == 1
-    assert first_event["spans"][0]["op"] == "cache"
-    assert first_event["spans"][0]["description"].startswith(
-        "get template.cache.some_identifier."
-    )
-    assert first_event["spans"][0]["data"] == {"cache.hit": False}
-
-    assert len(second_event["spans"]) == 1
-    assert second_event["spans"][0]["op"] == "cache"
-    assert second_event["spans"][0]["description"].startswith(
-        "get template.cache.some_identifier."
-    )
-    assert second_event["spans"][0]["data"] == {
-        "cache.hit": True,
-        "cache.item_size": 51,
-    }
-
-
-def test_cache_spans_disabled_templatetag(
-    sentry_init, client, settings, capture_events
-):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake",
-        }
-    }
-    sentry_init(
-        integrations=[
-            DjangoIntegration(
-                cache_spans=False,
-                middleware_spans=False,
-                signals_spans=False,
-            )
-        ],
-        traces_sample_rate=1.0,
-    )
-    events = capture_events()
-
-    client.get(reverse("view_with_cached_template_fragment"))
-    client.get(reverse("view_with_cached_template_fragment"))
-
-    (first_event, second_event) = events
-    assert len(first_event["spans"]) == 0
-    assert len(second_event["spans"]) == 0
-
-
 def test_csrf(sentry_init, client):
     """
     Assert that CSRF view decorator works even with the view wrapped in our own
